requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1

langchain==0.2.11
langchain-core==0.2.27
langchain-community==0.2.10
langchain-groq==0.1.5

google-search-results==2.4.2
beautifulsoup4==4.12.3
requests==2.32.3
tiktoken==0.7.0